+++ resources.go
@@ -92,8 +92,15 @@ func Provider() tfbridge.ProviderInfo {
 			},
 		},
 		Resources: map[string]*tfbridge.ResourceInfo{
-			"digitalocean_cdn":         {Tok: digitalOceanResource(digitalOceanMod, "Cdn")},
-			"digitalocean_certificate": {Tok: digitalOceanResource(digitalOceanMod, "Certificate")},
+			"digitalocean_cdn": {Tok: digitalOceanResource(digitalOceanMod, "Cdn")},
+			"digitalocean_certificate": {
+				Tok: digitalOceanResource(digitalOceanMod, "Certificate"),
+				Fields: map[string]*tfbridge.SchemaInfo{
+					"type": {
+						Type: digitalOceanType(digitalOceanMod, "CertificateType"),
+					},
+				},
+			},
 			"digitalocean_database_cluster": {
 				Tok: digitalOceanResource(digitalOceanMod, "DatabaseCluster"),
 				Fields: map[string]*tfbridge.SchemaInfo{
@@ -117,7 +124,14 @@ func Provider() tfbridge.ProviderInfo {
 					},
 				},
 			},
-			"digitalocean_droplet_snapshot":       {Tok: digitalOceanResource(digitalOceanMod, "DropletSnapshot")},
+			"digitalocean_droplet_snapshot": {
+				Tok: digitalOceanResource(digitalOceanMod, "DropletSnapshot"),
+				Fields: map[string]*tfbridge.SchemaInfo{
+					"regions": {
+						Elem: &tfbridge.SchemaInfo{Type: digitalOceanType(digitalOceanMod, "Region")},
+					},
+				},
+			},
 			"digitalocean_firewall":               {Tok: digitalOceanResource(digitalOceanMod, "Firewall")},
 			"digitalocean_floating_ip":            {Tok: digitalOceanResource(digitalOceanMod, "FloatingIp")},
 			"digitalocean_floating_ip_assignment": {Tok: digitalOceanResource(digitalOceanMod, "FloatingIpAssignment")},
@@ -129,17 +143,34 @@ func Provider() tfbridge.ProviderInfo {
 					},
 				},
 			},
-			"digitalocean_kubernetes_node_pool": {Tok: digitalOceanResource(digitalOceanMod, "KubernetesNodePool")},
+			"digitalocean_kubernetes_node_pool": {
+				Tok: digitalOceanResource(digitalOceanMod, "KubernetesNodePool"),
+				Fields: map[string]*tfbridge.SchemaInfo{
+					"size": {
+						Type: digitalOceanType(digitalOceanMod, "DropletSlug"),
+					},
+				},
+			},
 			"digitalocean_loadbalancer": {
 				Tok: digitalOceanResource(digitalOceanMod, "LoadBalancer"),
 				Fields: map[string]*tfbridge.SchemaInfo{
 					"region": {
 						Type: digitalOceanType(digitalOceanMod, "Region"),
 					},
+					"algorithm": {
+						Type: digitalOceanType(digitalOceanMod, "Algorithm"),
+					},
 				},
 			},
 			"digitalocean_project": {Tok: digitalOceanResource(digitalOceanMod, "Project")},
-			"digitalocean_record":  {Tok: digitalOceanResource(digitalOceanMod, "DnsRecord")},
+			"digitalocean_record": {
+				Tok: digitalOceanResource(digitalOceanMod, "DnsRecord"),
+				Fields: map[string]*tfbridge.SchemaInfo{
+					"type": {
+						Type: digitalOceanType(digitalOceanMod, "RecordType"),
+					},
+				},
+			},
 			"digitalocean_ssh_key": {Tok: digitalOceanResource(digitalOceanMod, "SshKey")},
 			"digitalocean_spaces_bucket": {
 				Tok: digitalOceanResource(digitalOceanMod, "SpacesBucket"),
@@ -156,10 +187,20 @@ func Provider() tfbridge.ProviderInfo {
 					"region": {
 						Type: digitalOceanType(digitalOceanMod, "Region"),
 					},
+					"initial_filesystem_type": {
+						Type: digitalOceanType(digitalOceanMod, "FilesystemType"),
+					},
 				},
 			},
 			"digitalocean_volume_attachment": {Tok: digitalOceanResource(digitalOceanMod, "VolumeAttachment")},
-			"digitalocean_volume_snapshot":   {Tok: digitalOceanResource(digitalOceanMod, "VolumeSnapshot")},
+			"digitalocean_volume_snapshot": {
+				Tok: digitalOceanResource(digitalOceanMod, "VolumeSnapshot"),
+				Fields: map[string]*tfbridge.SchemaInfo{
+					"regions": {
+						Elem: &tfbridge.SchemaInfo{Type: digitalOceanType(digitalOceanMod, "Region")},
+					},
+				},
+			},
 		},
 		DataSources: map[string]*tfbridge.DataSourceInfo{
 			"digitalocean_certificate":        {Tok: digitalOceanDataSource(digitalOceanMod, "getCertificate")},
@@ -188,9 +229,14 @@ func Provider() tfbridge.ProviderInfo {
 			},
 			Overlay: &tfbridge.OverlayInfo{
 				DestFiles: []string{
+					"certificateType.ts",
 					"databaseSlug.ts",
 					"dropletSlug.ts",
-					"region.ts", // Region union type and constants
+					"filesystemType.ts",
+					"loadbalancerAlgorithm.ts",
+					"protocol.ts",
+					"recordType.ts",
+					"region.ts",
 				},
 				Modules: map[string]*tfbridge.OverlayInfo{},
 			},