+++ sdk/nodejs/volumeSnapshot.ts
@@ -4,6 +4,8 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as utilities from "./utilities";
 
+import {Region} from "./index";
+
 /**
  * Provides a DigitalOcean Volume Snapshot which can be used to create a snapshot from an existing volume.
  * 
@@ -65,7 +67,7 @@ export class VolumeSnapshot extends pulumi.CustomResource {
     /**
      * A list of DigitalOcean region "slugs" indicating where the volume snapshot is available.
      */
-    public /*out*/ readonly regions!: pulumi.Output<string[]>;
+    public /*out*/ readonly regions!: pulumi.Output<Region[]>;
     /**
      * The billable size of the volume snapshot in gigabytes.
      */
@@ -128,7 +130,7 @@ export interface VolumeSnapshotState {
     /**
      * A list of DigitalOcean region "slugs" indicating where the volume snapshot is available.
      */
-    readonly regions?: pulumi.Input<pulumi.Input<string>[]>;
+    readonly regions?: pulumi.Input<pulumi.Input<Region>[]>;
     /**
      * The billable size of the volume snapshot in gigabytes.
      */